sched: cpufreq: sysfs knob for capacity_margin
This commit makes the capacity_margin knob configurable from user-space.
By tinkering with this knob, cpufreq scaling of frequency can be made
more aggressive. Allowed values for this knob range from 0 to 1024. By
default, it has the value of 256 which sets the frequency scaling factor
to be 1.25

Bug 2106745

Change-Id: I8bf427f116fc4956440315528ebcd8f820b839e5
Signed-off-by: Waqar Ali <waqara@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1758492
Reviewed-by: Bo Yan <byan@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Winnie Hsu <whsu@nvidia.com>
Reviewed-by: Bibek Basu <bbasu@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
Signed-off-by: DennySPB <dennyspb@gmail.com>