msm: mhi_dev: Update MHI MMIO initialization for BHI_INTVEC
BHI_INTVEC is programmed by the host and does not require AMSS
software to reprogram this register during initial boot.
With PCIe early initialization the host MHI driver can get
enumerated earlier and update this register and the existing
AMSS configuration for the register can overwrite this value.
Therefore remove this initialization.

Change-Id: Ie550db13060b509bf49895022674e6dd3ec0a8c2
Signed-off-by: Siddartha Mohanadoss <smohanad@codeaurora.org>
Signed-off-by: DennySPB <dennyspb@gmail.com>